(themes) refactor css variables to have more things theme-related:
- there is no more default values set in cssVars for vars being tied to
a theme, meaning `colors` and `vars` do not have any value anymore. They
just expose the new `tokens` object
- a new theme `tokens` object exposes all the CSS variables related to
the current theme. They include colors, but also font family, paddings,
basically everything that was described in cssVars colors and vars.
- a new theme `legacyVariables` object exposes all the previous cssVars
`theme` variables that are considered legacy code. The cssVars `theme`
object just consumes this new object
- themes now must define tokens that were before only in cssVars
`colors` and ` vars`. To help with that, there is now a Base.ts theme
that describes all variables that should stay the same for most themes.
- in practice, defining a new theme will mean meandescribing an object a
couple dozens variables, instead of the 400+ previously.
- Removed the use of ts-interface-builder as it didn't work with new
types (made with generics). the gristUrls code was updated accordingly.
But the AppModel that used the ThemePrefsChecker now misses the check. I
actually don't understand why it was actually necessary here. Removed it
for now, I need some feedback.

- example of "legacy variables" handling has begun: some more generic
token variables were added, that a set of legacy variables target
instead of previous specific values. it shows how we could try to
refactor the bunch of variables so that updating themes would then be
easier. The added variables are ones matching variables that were used a
lot in grist dark and grist light themes.