(core) Refactor ICreate/create methods in several Grist variants, replacing makeSimpleCreator with classes
Summary:
Previous approach using makeSimpleCreator() was confusing and repetitive. Now
each variant of Grist is configured with its own class, and customizations are
overrides of methods in that class. This should be easier to maintain going
forward.

Note that grist-desktop and grist-static will need follow-up refactors to match
this pattern.

Test Plan: No intended changes to behavior, relying on existing automated tests.

Reviewers: Spoffy

Reviewed By: Spoffy

Differential Revision: https://phab.getgrist.com/D4422