(core) Make matching of email for user-attribute tables case insensitive.
Summary:
The behavior of user.Email has been inconsistent: for requests via UI, it is
the potentially-capitalized email ("display email"), while for API requests
it's lowercase. Further, the display email isn't always knowable, as
invitations are case-insensitive, and the case-sensitive display email is set
at login time. For people with non-lowercase emails this has been a source of
problems in access rules.

This addresses the problem in two ways:
1. Matching of user.Email to User-Attribute tables is now case-insensitive.
2. The direct use of user.Email in rules (e.g. `user.Email == rec.Email`) is
    unaffected, to prevent breaking existing docs, but we now support
    predictable usage, using `user.Email.lower()`
    (e.g. `user.Email.lower() == rec.Email`).

This relies two new lower-level features:
- Predicate formulas (as used for ACL) now support functions and methods,
  currently limited to just .lower()/.upper() for strings.
- ServerQuery now accepts a custom SQL "where" clause (only ones constructed
  internally, not from outside), and can combine it with filters.

Some related fixes:
- For View-As users, avoid adding multiple options from user table that differ
  only in the case of email address.
- Fix cleanup of temp tables for DocStorage queries with too many parameters.
- Stricter checks for attribute lookups in evaluating access-rule conditions,
  to avoid exposing any unintended properties.

Test Plan:
Added tests for the new behavior, for the lower-level features, and
for the related fixes.

Reviewers: georgegevoian, paulfitz

Reviewed By: georgegevoian, paulfitz

Differential Revision: https://phab.getgrist.com/D4418