Install Chrome manually in workflows (#1398)
## Context

CI began experiencing failures after the version of Google Chrome used
by GitHub Actions workers was upgraded to 132. The root cause turned out
to be a problem with the version of Chrome for Testing downloaded by
Selenium Manager requiring the "--no-sandbox" argument to work.

Adding the argument revealed another problem: tests have become very
unreliable in recent versions of Chrome for Testing, which appears to
differ slightly from the regular version of Chrome. The root cause is
currently unknown.

## Proposed solution

Switch to manually installed versions of Chrome in workflows. Selenium
Manager will still download an appropriate version of chromedriver for
the installed version of Chrome.

## Has this been tested?

<!-- Put an `x` in the box that applies: -->

- [ ] 👍 yes, I added tests to the test suite
- [ ] 💭 no, because this PR is a draft and still needs work
- [x] 🙅 no, because this is not relevant here
- [ ] 🙋 no, because I need help <!-- Detail how we can help you -->