improve snapshot generation under load (#1380)
## Context

Grist uses the SQLite API to make backups. This PR reorganizes the code
so that the SQLite connection used by ActiveDoc (the main class
representing an open document) is also the one used for backups.
Otherwise, if the document is under continuous modification (e.g.
because of some automation), attempting to perform the backup may not
terminate since SQLite may keep restarting the backup under the hood.

With this change, it is now also worth beginning the backup even if the
document is busy, so the PR also does that.

In testing, when a document is under heavy load, the last step call in
the SQLite backup API may be relatively slow, perhaps as all changes
that accumulated are flushed also to the backup. But since a since
SQLite connection is being used now, there should not be any new
SQLite-level busy timeouts.

## Has this been tested?

<!-- Put an `x` in the box that applies: -->

- [x] 👍 yes, I added tests to the test suite
- [ ] 💭 no, because this PR is a draft and still needs work
- [ ] 🙅 no, because this is not relevant here
- [ ] 🙋 no, because I need help <!-- Detail how we can help you -->