(core) exclude the last step of backups from stats used to assess disk health
Summary:
When producting a snapshot of an active Grist doc, the first step
is to make a local backup of it that will be stable during a copy.
The SQLite backup API operates incrementally, with each step copying
up to a bounded number of pages. The last step does extra work, if
the backup is done using the connection to the source as opposed to
the destination, as we now do. We measure the maximum time taken by
steps during a backup, which proved to be a convenient measure of
disk health. But the variation in the last step disrupts that usage.
This commit provides a measurement that excludes the last step of
backup.

For example, a test run gave on a backup gave:

 * finalStepTimeMs=859
 * maxStepTimeMs=859
 * maxNonFinalStepTimeMs=11
 * numSteps=161

The maxNonFinalStepTimeMs measurement excludes the final step,
now reported separately.

Test Plan: log values inspected on runs of backupSqliteDatabase test

Reviewers: georgegevoian

Reviewed By: georgegevoian

Differential Revision: https://phab.getgrist.com/D4441