(themes) change theme loading for better grist-plugin-api support
In order to prevent "context switching" when dealing with theme
variables, choice was made to now stick with js-like keys in theme
files. So that consuming tokens in the app, and writing tokens in
themes, would be with the same keys (like 'bgSecondary' everywhere
instead of 'bgSecondary' in one part and 'bg-secondary' in another).

This initialy changed how theme css vars were attached to the DOM.

However it implied modifying how external plugins load the theme
variables. This is actually something we want to avoid, because writing
breaking changes in the grist-plugin-api file is rather annoying to
support.

This commit rollbacks the way theme vars were attached so that it
works as before, meaning, the theme loading function gets a list of "css
like" names instead of "js like" names.