specify NODE_ENV=production by default in docker image
The express web framework that Grist uses is sensitive to a NODE_ENV
environent variable. Grist also gives some operations more time when
this variable is set. At Grist Labs we have this flag set for our
SaaS, but until now it hasn't been set in the regular docker image.