(kb) allow classic keyboard nav when not viewing a document
1) Disable the clipboard catch-all element when there is no need to
enable it (we conclude we don't need it when we see no copy/put/paste
command is set).

This means it is disabled on most pages except the document view.

2) Make sure interactive elements are highlighted when focusing them
with the keyboard. Lots of elements have `outline: none` forced in the
codebase to prevent showing unwanted outlines when clicking with the
mouse. But it prevents keyboard users to see where they are. Have a tiny
JS script checking for keyboard interactions to force showing the
outline on actual kb focus.