(kb) introduce "regions" keyboard navigation with RegionFocusSwitcher
RegionFocusSwitcher makes it possible for the user to cycle through app
panels and document widgets (named "regions" as a whole) with a couple
keyboard shortcuts.

1) On a grist document page, pressing the Ctrl+o (or Ctrl+Shift+o)
shortcut jumps your keyboard focus from widget to widget, but also panel
to panel. You go to widget1, widget 2, then left panel, top panel, then
widget 1, widget 2, etc. This shortcut is necessary to reach stuff
outside widgets: because by default, usual kb shortcuts like tab or
arrow keys are highjacked by widgets and you can't go out ot them.

On other pages (that don't have widgets), pressing the shortcut jumps
your keyboard focus from panel to panel, but this time it also includes
the main content area as a "panel". The shortcut is not _that_ necessary
on other pages since you can now use Tab normally. But I guess it's
better to keep the shortcut always work whatever the page.

2) The new creatorPanel command allows to keyboard focus from/to the
creator panel with Ctrl+Alt+o.
This is a separate command from the regions-cycling, as the creator
panel is tied to the currently focused viewLayout section. So we must be
able to focus a section, and then directly go to the creator panel in
order to configure it.

If we notice the user just spammed the Ctrl+O shortcut too much, we
assume he just tried accessing the creator panel and we show him the
keyboard shortcut in a notification.

3) Pressing Esc while on a panel will first focus back the panel, then
restore focus to the view layout.

4) The grist doc view now keeps track of this new stuff, in order to
disable the view section commands when focusing panels. This is an
important trick that lets us enable the Tab key and others when focusing
the panels.

5) The new navigation shortcuts are now always on, even in inputs.
Using ctrl+o is somewhat problematic because it is the default browser
shortcut for opening files, so in cases where we don't catch the
keypress, the filechooser appears. We want to avoid that as much as
possible.

6) One main trick allowing us to easily allow Tab navigation on panels,
while on a grist docs page, is the change in Clipboard.js. We can now
set a 'clipboard_group_focus' class somewhere in the dom, and the
clipboard will allow focusing elements inside the one having the class.

6) On a grist doc, mouse interactions also drastically impact keyboard
navigation: they now silently toggle the view layout commands. If I
click on the left panel, it wont actually force a focus change, but it
will enable keyboard nav inside the panel.

This is not done yet:
- needs more actual usage testing to finetune the behavior and make sure
it is what we want.
- needs tests