(core) Implementing virtual tables without GristDoc
Summary:
Removing dependency on live GristDoc for virtual tables (and Grid component).
Adding a wrapper for ExternalTables that mocks GristDoc and DocData that allows
showing GridView component on pages that don't have access to GristDoc.

To test it, TimingPage was refactor to use the new wrapper.

Test Plan: Existing test for Timing page.

Reviewers: georgegevoian

Reviewed By: georgegevoian

Subscribers: dsagal

Differential Revision: https://phab.getgrist.com/D4438