(themes) Handle custom.css after theme variables refactoring
New implementation of themes variable handling (previous commit) broke
the support of custom.css files that don't use the new variables.

This issue is mentioned by George and explained in details as to why in
my comment:
https://github.com/gristlabs/grist-core/pull/1340#issuecomment-2662147743

This commit fixes that.
Example custom.css file was also updated to use the new variables, and
explain how to migrate.

Example:

1) Variables on load looks like this (in grist app):

```css
:root {
  --grist-color-dark: var(--grist-theme-body);
  --grist-theme-body: #262633;
}
```

2) I have an "old" custom.css file (as a selfhoster), it contains this:

```css
:root {
    --grist-color-dark: #ff00ff !important;
}
```

Without this commit, the UI breaks because the link to the new var is
gone. I would need to manually update my custom.css file to change
`--grist-theme-body` instead.

With this commit, Grist appends this to the DOM automatically:

```css
  :root {
    ---grist-theme-body: var(--grist-color-dark) !important;
  }
```

It also adds a warning in the console in order to encourage devs to
update their custom.css file.