refine the interaction between backups and main document connections
Recently we changed to making backups using the main document
connection, when available. The system making the backups and
the system managing the document are quite loosely coupled, so
the main document connection can get closed at any time. The
backup process was written to expect this, but a problem appeared
on the closing side: SQLite will refuse to close if a backup is
in progress, and that error was entirely uncaught due to a missing
await.

This adds the needed await. It also adds some explicit robustness
on both sides so that it is easier to tell from local considerations
that shutdown should work.