refine the interaction between backups and main document connections (#1461)
## Context

Recently we changed backups to use the main document connection, when
available, since then they can complete even while the document is
changing constantly. The system making the backups and the system
managing the document are quite loosely coupled, so the main document
connection can get closed at any time, even during a backup. The backup
process was written to expect this, but a problem appeared on the
closing side: SQLite will refuse to close if a backup is in progress,
and that error was entirely uncaught due to a missing await.

## Proposed solution

This adds the needed await. It also adds some explicit robustness on
both sides

## Related issues

https://github.com/gristlabs/grist-core/pull/1380
https://github.com/gristlabs/grist-core/issues/1427

## Has this been tested?

- [x] 👍 yes, I added tests to the test suite
- [ ] 💭 no, because this PR is a draft and still needs work
- [ ] 🙅 no, because this is not relevant here
- [ ] 🙋 no, because I need help <!-- Detail how we can help you -->