Keyboard shortcuts page dismissal fix (#1473)
You can use `⌘ + /`, `ESC`,or click outside the popup window to dismiss
it.

## Context

Updated the keyboard shortcuts popup code so it would hide/dismiss the
popup by setting `stye="display:none;"` on the div with
`class="g-help"`. Now when a user clicks outside of the popup window,
hits the `ESC key`, or toggles with `⌘ + /` (I didn't test `Ctrl + /` as
I don't have windows) the window is dismissed. Previously the popup
window would not hide/dismiss as outlined in the issue below.

NOTE: ESC still does not work on Safari.

I'm pretty much just a hack coder so please review. I started using
Grist a couple of months back and just noted this behaviour a couple of
days ago. I saw the related issue and thought I'd try and fix. The popup
is now working without issue for me. This is my first pull request for
the project so use at your discretion.

## Related issues

https://github.com/gristlabs/grist-core/issues/796

## Has this been tested?

I tested on Firefox, Chrome, and Safari, on MacOS Sonoma. No issues with
Chrome, or Firefox. On Safari the ESC key did not hide/dismiss the popup
window, but clicking outside the window or cycling with `⌘ + /` still
worked.


https://github.com/gristlabs/grist-core/pull/1473