Adds zip downloads for all of a document's attachments (#1471)
## Context

When using internal storage for attachments, all attachments are
included in the .grist file when a document is downloaded. They can be
extracted if needed using standard SQLite tools.

When using external storage, attachments are stored outside of Grist
(e.g. MinIO, S3, filesystem) and can't be easily retrieved.

## Solution

This PR adds an API endpoint to download a .zip of all attachments in
the document.

This is streamed where possible, resulting in no more than a single file
being kept in memory at once (due to SQLite restrictions - external
storage keeps no copies in memory).

## Related issues

#1021