support more options for SQLite connection
The basic connection parameters used to open documents hasn't
been touched in a long time. This adds support for WAL mode,
or default mode with SYNCHRONOUS=full.

Use `GRIST_SQLITE_MODE=wal` or `GRIST_SQLITE_MODE=sync` for this
new behavior. Default is unchanged.

WAL mode is not recommended for Desktop, or similar situations
where you may habitually interact with .grist files managed
by the Grist application on the file system. If all interaction
is via the API, it should be a fine mode to use.

The sync mode is a candidate for use on the Desktop.