(tests) add docs helping people with the current e2e test issues
We noticed recently that e2e tests don't work with Chromium (the Chrome
for Testing binaries downloaded by selenium webdriver), and with Chrome
builds starting with version 134.

While we should work on fixing the actual tests so that everything works
correctly on Chrome for Testing, as for now, a workaround is to also use
a pinned _Chrome_ version locally.

This adds some documentation to help people understand the issue and act
accordingly.