Adds support for downloading attachments as a .tar (#1495)
Currently only Zip is supported for downloading attachments. In the
future, we want to be able to support re-uploading attachments. Zip will
make this difficult, as the structure of the zip file is stored at the
end of file, requiring the entire file to be uploaded before it can be
parsed.

This commit adds downloading attachments as a .tar, which uses inline
headers for each file it contains. This means it can be easily parsed
and processed through node's streams, minimizing server disk and memory
usage.

## Related issues

#1021