(core) Fix bug with saving access rules that use "in" operator with lists of choices and similar
Summary:
This includes a change in behavior. Previously predicate like `NEEDLE in HAYSTACK` would work as expected when `HAYSTACK` were a list or a string, and would return false when it were null or undefined. For other values (e.g. numbers or references) it would fail with an error ("t.includes is not a function") during evaluation.

A recent change that added some minimal type-checking exposed the same error during saving of rules ("t.includes is not a function"), interfering with all rules of this form.

Now using `in` with wrong types will return false, rather than fail. (Similarly, "not in" will return true.)

Note a subtle difference for wrong types: previously, the rule would raise an exception, and access rules would try to interpret it in the more restrictive way (e.g. "deny access when `user.Email in 5`" would be interpreted as "deny"); now the condition would return a boolean value, and the access would be allowed or denied accordingly (e.g. "deny access when `user.Email in 5`" would not match, and would not deny access).

Test Plan: Updated test cases for the changed behavior.

Reviewers: georgegevoian, paulfitz

Reviewed By: georgegevoian, paulfitz

Subscribers: paulfitz

Differential Revision: https://phab.getgrist.com/D4487