(core) Set aclAsUser param in same document URLs
Summary:
URLs pointing to the same document are normally opened in the same
tab without reload. However, when viewing a document as a different
user, these URLs were opening in new tabs because the aclAsUser param
was not included in any URLs generated by Grist (e.g. `SELF_HYPERLINK()`),
and two URLs with different link parameters are considered not to be
same document URLs.

Now, we automatically set aclAsUser (when present) in any URLs pointing
to the same document, restoring the expected behavior.

Test Plan: Browser tests.

Reviewers: dsagal

Reviewed By: dsagal

Subscribers: dsagal

Differential Revision: https://phab.getgrist.com/D4489