choice tokens: change bg/text shade automatically to improve readability
Choice token text or background, when not explicitely overriden by the
user, change depending on light/dark theme.

For now, this can be an issue when a custom background color is set,
while text color is not, or vice versa.

This is an attempt at fixing that. The idea is to automatically switch
to light or dark text when only the bg is user-defined, depending on the
defined bg. And act the same if only the text is user-defined.