(core) Include documentation link in summary table error
Summary:
The error toast shown when adding a non-formula column to a summary
table didn't contain a link to the documentation. Instead, there was a
link to contact support directly for help. That's now been replaced
with a link to the documentation, which explains how to properly use
summary tables.

Test Plan: Manual.

Reviewers: jarek

Reviewed By: jarek

Subscribers: jarek

Differential Revision: https://phab.getgrist.com/D4494