forms: have labels tied to inputs for better screen reader support
- make every input correctly be tied to their labels by adding `id`
attributes. This helps screen reader users a lot

- for radio groups and checkbox groups, simulate fieldset/legend with
`role=group` + `aria-labelledby` attributes on the containers*. We
simulate with aria roles to prevent css issues and having to handle
those changes in current code. The result is the same for screen
readers.

- markup is not the cleanest as is, because now radio/checkbox groups
still use a `label` tag for their heading, that has an `id` attr
pointing to nowhere. It's not an issue in real use though. It makes the
js code cleaner to not handle the switch to another dom el/removing the
for attr.