George's work on allocating docs to workers
Documents have previously been assigned to workers randomly, without accounting
for resource utilization of available workers. Now, documents are assigned to workers
deterministically by attributing a score (0.0 to 1.0, inclusive) to each worker, and picking
the worker with the highest score. Scores are a computation of the ratio of used to total
system memory (as reported by the OS), with some adjustments made for documents
currently in flux (i.e. in the middle of loading) or assigned but not yet opened.

Worker state such as memory usage is maintained by DocWorkerMap, which is backed by
Redis in multi-worker environments. Memory usage is updated on interval (~5 seconds), and
whenever the state of a document changes (e.g. starts or finishes loading).