OIDC and SAML environment variables (#1527)
Added reference to all environment variables linked to OpenID and SAML
configuration with a link to the documentation

## Context

After a discussion on [Grist Community
forum](https://community.getgrist.com/t/oidcconfig-email-not-verified/9032),
it has been suggested to update the list of environment variables.

## Proposed solution

All environment variables linked to OpenID and SAML are missing in the
readme file, I propose to add two lines referencing prefix GRIST_OIDC
and GRIST_SAML respectively, and provide a link to the complete
documentation with the full list for each.

## Related issues

None

## Has this been tested?

<!-- Put an `x` in the box that applies: -->

- [ ] 👍 yes, I added tests to the test suite
- [ ] 💭 no, because this PR is a draft and still needs work
- [x] 🙅 no, because this is not relevant here (update of readme file)
- [ ] 🙋 no, because I need help <!-- Detail how we can help you -->