Adds archive uploading to add missing attachments (#1506)
## Context

When a document using external attachments is downloaded then uploaded,
none of the attachments will work. This is due to external attachments
being stored on a per-document basis.

## In this commit

This commit allows any missing files to be populated in the external store,
by uploading a .tar archive containing the missing attachments.

It only supports .tar archives downloaded from Grist, as it uses the
file's name to check if the file is missing.

## Related issues

#1021