prepare a Contributor License Agreement (#1546)
This adds a Contributor License Agreement for use in a workflow for
people making pull requests.

The Apache 2.0 license that applies to the repo includes a mini-CLA
within it, but it is nonetheless important to have an affirmatively
signed CLA. Models: [TypeScript
repo](https://github.com/Microsoft/TypeScript/blob/main/CONTRIBUTING.md#legal),
[gVisor
repo](https://github.com/google/gvisor/blob/master/CONTRIBUTING.md#contributor-license-agreement).

This PR does not introduce the workflow. That will be follow-up that is
more straightforward to prepare once this document is in place.