Enhance build.sh to allow building using Webpack in dev mode (#1545)
## Context

Since fee42ae, `yarn build:prod` is meant to build the app in production
mode (slower process) and `yarn build` in dev mode (faster).

But the mode has not been passed to webpack.

## Proposed solution

We either pass `--mode production` or `--mode development` to webpack
depending on the arguments passed to the script.

With this commit, building in dev mode takes less than 10s (previously
approx. 1 min).

## Related issues

None

## Has this been tested?

<!-- Put an `x` in the box that applies: -->

- [ ] 👍 yes, I added tests to the test suite
- [ ] 💭 no, because this PR is a draft and still needs work
- [x] 🙅 no, because this is not relevant here
- [ ] 🙋 no, because I need help <!-- Detail how we can help you -->