FlexServer: add a `publishLatestVersionAvailable` method
if Redis is available, we'll use it for pubsub and make sure that all
home servers store the latest version in their
`_latestVersionAvailable` member.

If it's not available, we just skip the publishing and directly save
the variable on the current home server.

The subscription will come in the next commit.