Adds /docs/import endpoint, makes `/workspaces/{wid}/import` support files (#1535)
## Context

The current API for copying and uploading Grist documents is `POST
/api/docs`.

However, this method is heavily overloaded, doing all of:
- Copying an existing document
- Uploading a new document from a file
- Creating a new saved document
- Creating a new unsaved document

This makes it tricky to document and use.

## Solution

This PR adds a new endpoint:
- `POST /api/docs/{docId}/copy`

And updates `POST /api/workspaces/{wid}/import` to support file uploads.

These accomplish the equivalent of `POST /api/docs`'s copy and import
behaviour, respectively.