add "yarn cli sqlite query" utility (#1547)
Add a utility for runing a query on an SQLite file and passing the
result through standard Grist decoding for BLOBs. This is handy for
reading objects of the wrong type stored as BLOBs to protect them in a
"cell", or for reading action history, e.g.

```sh
  yarn --silent cli sqlite query --json doc.grist \
    "select * from _gristsys_ActionHistory order by actionNum"
```

This will return the same content as the query would from the SQLite3
utility, except BLOBs will be expanded.