(core) SMTPNotifier: new SMTP-based notifier
Summary:
This adds a new notifier based on Nodemailer. It can be configured by
passing in a JSON configuration via the `GRIST_NODEMAILER_CONFIG`
environment variable, which the user can use to configure a Nodemailer
transport exactly as the Nodemailer configuration allows. Since
Nodemailer allows for much flexibility in configuration, it seems best
to let the user set the exact same JSON that Nodemailer would require.

Note that unlike Sendgrid, which externally was once configured via
`config.json`, we chose the environment variable for configuration for
a few reasons:

  1. External attachments already set a precedent for putting JSON
     config in an env var.

  2. Multi-server setups would require propagation of a `config.json`
     file, whereas an environment variable is easier to replicate
     (hence why external attachments already made the choice to supply
     a JSON configuration via an envvar)

  3. The `config.json` configuration for Sendgrid is actually
     currently broken as of July 8, 2024 when we changed the type of
     `gristServer.config` to `IGristCoreConfig`. That's a big problem,
     but a problem for another time.

For now, all of the notification email templates are hardcoded in the
repository.

Test Plan:
Manually set up by setting `GRIST_NODEMAILER_CONFIG` and try adding
users to a document. Nodemailer recommends using
[Ethereal](https://ethereal.email) as a test SMTP server.

Reviewers: georgegevoian

Reviewed By: georgegevoian

Differential Revision: https://phab.getgrist.com/D4538