(kb) differenciate global view commands from focused-only commands
sometimes, we want to run "current view"-related commands via keyboard
shortcuts, while we are focusing ie. the creator panel.

this commit makes it possible, by distinguishing "globally active"
view commands from "only user-focused" view commands.

in order to have the "globally active" commands more reliably work, we
make sure the active section is focused by the region switcher
automatically. it's not an ideal behavior for the user, as he will then
need to manually go focus back the panel he was on ; but this is more
done as a safety measure to not end up in a state where `body` is
focused while we'd expect otherwise.