Small improvements on forms accessibility (#1536)
## Context

Here are small HTML changes on the user-generated forms, mostly made to
improve the life of screen reader users.

## Proposed solution

Each fix has his own contained commit with its description.

The most important fix is the one making sure form labels are tied to
their inputs, as this makes screen reader navigation in forms much
easier.

Other fixes are arguably not _that_ important, but each improves the UX
a tiny bit for screen reader users, making this PR as a whole a
noticeable change.

## Related issues

This doesn't fix everything. As forms are created by users, what someone
does when authoring the form can have much impact on its accessibility.
For example, the form widget name or the headings hierarchy need to be
authored with care. We could imagine improving UX when creating forms to
make people more aware of that.

One thing that would be great in the future would also to define the
`lang` attribute on the html tag depending on the form locale. This
would make sure that screen readers spell words in the correct language.
This one thing is easy to code if relying on the document locale, but I
fear that applying that as of _right now_ might do more harm than good,
because of how document locales are currently handled in Grist. So I'm
more inclined to not change anything right now on this, this will
require some specific work.

## Has this been tested?


- [x] 👍 yes, I added tests to the test suite