onRecord: add an `expandRefs` boolean option
This is named after the similar `expand_refs` kwarg of the `RECORD`
function. By default and for backwards compatibility, `expandRefs` is
`true`. It can be set to `false` to return the reference IDs
themselves instead of the dereferenced values.

Note that for a single column of type `Reference`, this returns an object such as

  {
    tableId: 'TheReferencedTable',
    rowId: 2
  }

whereas for a `ReferenceList` column, it returns a list of IDs such as `[1, 2, 3]`.