(kb) regionFocusSwitcher internals: keep everything in state
not sure this is a good way to handle that, but it was a bit convoluted
to keep track of state change initiator in "_currentUpdateInitiator".
Now that it's in the state object it's more logical I think.

But i'm not sure of the implications of putting a mouse event object in
an observable?