(core) Add new AI assistant
Summary:
A new version of the AI assistant (V2) has been added, replacing the
previous version (now renamed to V1). This version of the assistant is
currently only used by the SaaS and Enterprise flavors of Grist. In
addition to supporting formula assistance, it can add, update, or delete
records, assist with trigger formulas, and answer queries about the
structure and data of a document.

Depending on which version of the assistant is used by the server, the
UI will show one of two things:
 - For version 1, the AI assistant panel in the formula editor
 - For version 2, a new AI Assistant button in the left panel

Test Plan: Browser and server tests.

Reviewers: paulfitz, jarek

Reviewed By: paulfitz, jarek

Differential Revision: https://phab.getgrist.com/D4541