add a confirmation modal to all DELETE operations on API console (#1583)
The API console page has a lot of endpoints on it, straight from they
grist.yml specification, and some of them could do a lot of harm if you
don't know what they are. This adds a confirmation modal to the DELETE
operations.

## Context

A user clicked through all the steps of doing an org deletion, without
realizing it would delete their team site. That endpoint isn't great,
and needs better documentation, but also it would be good to flag
deletions as something you should only do if you're confident in your
understanding.